// src/types/database.types.ts
export type Database = {
// ...
  public: {
    Tables: {
// ...
      work_logs: {
        Row: {
          created_at: string;
          custom_rate: number | null;
          date: string | null;
          default_rate: boolean | null;
          end_time: string | null;
          id: number;
          notes: string | null;
          organisation_id: string | null;
          paid: boolean;
          start_time: string | null;
          updated_at: string;
          user_id: string | null;
        };
        Insert: {
          created_at?: string;
          custom_rate?: number | null;
          date?: string | null;
          default_rate?: boolean | null;
          end_time?: string | null;
          id?: number;
          notes?: string | null;
          organisation_id?: string | null;
          paid?: boolean;
          start_time?: string | null;
          updated_at?: string;
          user_id?: string | null;
        };
        Update: {
          created_at?: string;
          custom_rate?: number | null;
          date?: string | null;
          default_rate?: boolean | null;
          end_time?: string | null;
          id?: number;
          notes?: string | null;
          organisation_id?: string | null;
          paid?: boolean;
          start_time?: string | null;
          updated_at?: string;
          user_id?: string | null;
        };
        Relationships: [
          {
            foreignKeyName: "work_logs_organisation_id_fkey";
            columns: ["organisation_id"];
            isOneToOne: false;
            referencedRelation: "organisations";
            referencedColumns: ["id"];
          },
        ];
      };
    };
// ...
  };
};